updated collision checking

+++ inc/Hvg.h
@@ -25,7 +25,7 @@ public:
     //new functions
     void scan(shared_ptr<HVGNode> node, Env<StateXY>* e);
     set<StateXY> getVG(HVGNode node);
-    int shortPathFromVG(set<StateXY> vg, StateXY start, StateXY goal);
+    double shortPathFromVG(set<StateXY> vg, StateXY start, StateXY goal);
     vector<shared_ptr<HVGNode>> getChildren(shared_ptr<HVGNode> parentNode, StateXY& parentState, Env<StateXY>* env);
     StateXY start;
